% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_number_by.R
\name{any_by}
\alias{any_by}
\title{Returns a logical TRUE if one row is TRUE}
\usage{
any_by(df, x, by = NULL)
}
\arguments{
\item{x}{a logical vector}

\item{by}{name(s) of variable(s) which determines groups}
}
\value{
an vector of logical
}
\description{
Returns a logical TRUE if one row is TRUE
}
