% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_when_mult.R
\name{case_when_mult}
\alias{case_when_mult}
\title{Case_when like function with multiple constant values}
\usage{
case_when_mult(...)
}
\arguments{
\item{...}{A sequence of two-sided formulas. The left hand side (LHS)
  determines which values match this case. The right hand side (RHS)
  provides the replacement value.

  The LHS must evaluate to a logical vector. Each logical vector can
  either have length 1 or a common length. All RHSs must have the same length,
  generally list or vector of constant values}
}
\value{
A vector when only one-length value ; otherwise a data.table with
  one column per value
}
\description{
Case_when like function with multiple constant values
}
\examples{
library(data.table)
dt <- data.table(a = c(1:5, NA, 7:10), b=rnorm(10), c=letters[1:10])
dt[, c("w", "x") :=
  case_when_mult(
    a \%\% 4 == 0 ~ list("p", 1.5),
    a \%\% 2 == 0 ~ list("m", 4) ,
    T ~ list("i", 6)
  )
]
print(dt)
}
