% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_by.R
\name{coalesce_by}
\alias{coalesce_by}
\title{coalesce_by Replace NA by constant, previous or next value, optionally by group}
\usage{
coalesce_by(dt, by, var = NULL)
}
\arguments{
\item{dt}{a data.table}

\item{by}{name(s) of variable(s) which determines groups}

\item{var}{name(s) of variable(s) with atomic values ; if 'var' is missing, all variables not in 'by' are selected}
}
\value{
a named list with item for each var
}
\description{
List of
}
\examples{
library(data.table)
ngrp <- 18000
nbygrp <- 50
dt <- data.table(
  id=rep(1:ngrp, each=nbygrp) * 2,
  a=sample(1:100, nbygrp * ngrp, replace=TRUE),
  b=sample(LETTERS, nbygrp * ngrp, replace=TRUE),
  c=sample((1:100 + 0.5), nbygrp * ngrp, replace=TRUE),
  d=sample(1:100, nbygrp * ngrp, replace=TRUE),
  e=sample(1:100, nbygrp * ngrp, replace=TRUE)
)
dt[(1:.N \%\% 2 == 1), a:=NA]
dt[(1:.N \%\% 3 < 2), b:=NA]
dt[(1:.N \%\% 4 < 3), c:=NA]
dt[(1:.N \%\% 5 < 4), d:=NA]
dt[(1:.N \%\% 6 < 5), e:=NA]

coalesce_by(dt, "id")
vars = c("b", "c", "e")
coalesce_by(dt, var=vars, by=c("id", "a"))
}
