% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumsum_by.R
\name{cummax_by}
\alias{cummax_by}
\title{Cumulative max by}
\usage{
cummax_by(dt, var = NULL, by = NULL)
}
\arguments{
\item{dt}{a data.table}

\item{var}{name(s) of variable(s) with atomic values ; if 'var' is omitted, all variables not in 'by' are selected}

\item{by}{name(s) of variable(s) which determines groups (optional) ; if 'by' is omitted, dt is considered as one group}
}
\value{
a list with item for each var
}
\description{
Cumulative max by
}
