% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtutils-package.R
\docType{package}
\name{dtutils-package}
\alias{dtutils}
\alias{dtutils-package}
\title{dtutils: Convenience Functions for the R Package data.table}
\description{
dtutils provides convenience functions for the R package data.table.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/and3k/dtutils}
}

}
\author{
\strong{Maintainer}: Bela Hausmann \email{belahausmann@fsfe.org} (\href{https://orcid.org/0000-0002-0846-1202}{ORCID})

}
\keyword{internal}
