% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_groupingsets.R
\name{gen_groupingsets}
\alias{gen_groupingsets}
\title{Helper for creating groupingsets}
\usage{
gen_groupingsets(dim, prefix = 0)
}
\arguments{
\item{dim:}{character vector containing variable's names}

\item{prefix:}{number of \emph{non-crossing} first variables}
}
\description{
Helper for creating groupingsets
}
\examples{
> gen_groupingsets(c("a", "b", "c", "d", "e"), prefix=3)
[[1]]
[1] "a" "b" "c"

[[2]]
[1] "a" "b" "c" "e"

[[3]]
[1] "a" "b" "c" "d"

[[4]]
[1] "a" "b" "c" "d" "e"
}
