% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_fill_by.R
\name{group}
\alias{group}
\title{Group by variable}
\usage{
group(dt, by)
}
\arguments{
\item{dt}{a data.table}

\item{by}{name(s) of variable(s) which determines groups}
}
\value{
a forderv vector with 'starts' attribute
}
\description{
Return a forderv vector
}
