% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tsv.R
\name{internal_write_tsv}
\alias{internal_write_tsv}
\title{Internal wrapper for \code{fwrite}}
\usage{
internal_write_tsv(object, file, header = TRUE)
}
\arguments{
\item{object}{An \R object to save as TSV.}

\item{file}{Output file name.
File extensions other than \code{.tsv}, \code{.tab}, and \code{.txt} result in a warning.}

\item{header}{Should the column names be written?}
}
\description{
Calls \code{\link[data.table:fwrite]{data.table::fwrite()}} with appropriate arguments:
\itemize{
\item The column separator is always \code{"\\t"}.
\item The line separator is always \code{"\\n"},
independent of the operating system.
}
}
