% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_number_by.R
\name{last_by}
\alias{last_by}
\title{Returns a logical TRUE for last row of each group}
\usage{
last_by(dt, by = NULL)
}
\arguments{
\item{dt}{a data.table}

\item{by}{name(s) of variable(s) which determines groups}
}
\value{
an vector of integer
}
\description{
Returns a logical TRUE for last row of each group
}
\examples{
dt <- data.table(id = rep(1:10, each=5))
dt[first_by(dt, "id"), pos := "F"]
dt[last_by(dt, "id"), pos := "L"]
}
