% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_fill_by.R
\name{na_fill_by}
\alias{na_fill_by}
\title{Replace NA by constant, previous or next value, optionally by group}
\usage{
na_fill_by(dt)
}
\description{
Replace NA by constant, previous or next value, optionally by group
}
