% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_fill_by.R
\name{na_replace}
\alias{na_replace}
\title{Replace NA by constant}
\usage{
na_replace(dt, fill, inplace = FALSE)
}
\arguments{
\item{dt}{a data.table}

\item{fill}{NA is replaced by this 'fill' value}

\item{inplace}{when inplace = TRUE, na_fill is compute inplace}
}
\value{
a list with item for each var
}
\description{
Replace NA by constant
}
\note{
All NA values are replaced by the 'fill' value. Ex with 'fill = 0' : \code{c(NA, 1, NA, NA, 2, NA, 3, NA)}
gives \code{c(0, 1, 0, 0, 2, 0, 3, 0)}
}
