% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_number_by.R
\name{row_number_by}
\alias{row_number_by}
\title{Create a sequence 1...N by group}
\usage{
row_number_by(dt, by)
}
\arguments{
\item{dt}{a data.table}

\item{by}{name(s) of variable(s) which determines groups}
}
\value{
an vector of integer
}
\description{
Create a sequence 1...N by group
}
\examples{
dt1 <- data.table(id = rep(1:10, each=5))
dt1[, row_num := row_number_by(dt, "id")]
dt2 <- data.table(id1 = rep(1:100, each=50), id2 = rep(1:1000, each=5))
dt2[, row_num := row_number_by(dt, c("id1", "id2"))]
}
