% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_by.R
\name{shift_by}
\alias{shift_by}
\title{shift/lag by group}
\usage{
shift_by(dt, var = NULL, by = NULL, inplace = FALSE)
}
\arguments{
\item{dt}{a data.table}

\item{var}{name(s) of variable(s) with atomic values ; if 'var' is omitted, all variables not in 'by' are selected}

\item{by}{name(s) of variable(s) which determines groups (optional) ; if 'by' is ommitted, dt is considered as one group}

\item{inplace}{when inplace = TRUE, na_fill is compute inplace}
}
\value{
a list with item for each var
}
\description{
shift/lag by group
}
