% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{weightedsum}
\alias{weightedsum}
\title{Compute weigthed sum}
\usage{
weightedsum(dt, wt, by = NULL, na.rm = F)
}
\arguments{
\item{dt}{a data.table}

\item{wt}{numeric vector of weights}

\item{by}{name(s) of variable(s) which determines groups}
}
\value{
A dataframe
}
\description{
Compute weigthed sum
}
