% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{wsum_by}
\alias{wsum_by}
\title{Compute weighted sum}
\usage{
wsum_by(dt, wt, by = NULL, na.rm = F)
}
\arguments{
\item{dt}{a data.table}

\item{wt}{numeric vector of weights}

\item{by}{name(s) of variable(s) which determines groups}
}
\value{
A dataframe
}
\description{
Compute weighted sum
}
