% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{DT}
\alias{DT}
\title{wrapper of [}
\usage{
DT(
  x,
  i,
  j,
  by,
  keyby,
  with = TRUE,
  nomatch = NA,
  mult = "all",
  roll = FALSE,
  rollends = if (roll == "nearest") c(TRUE, TRUE) else if (roll >= 0) c(FALSE, TRUE) else
    c(TRUE, FALSE),
  which = FALSE,
  .SDcols,
  verbose = getOption("datatable.verbose"),
  allow.cartesian = getOption("datatable.allow.cartesian"),
  drop = NULL,
  on = NULL,
  env = NULL
)
}
\description{
wrapper of [
}
\examples{
dt <- data.table(a = c(1:5, na, 7:10), b=rnorm(10), c=letters[1:10])

DT(dt, a \%\% 2 == 0, .(b, d=2*b))
}
