% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_fill_by.R
\name{na_fill_by}
\alias{na_fill_by}
\title{Replace NA by constant, previous or next value, optionally by group}
\usage{
na_fill_by(dt, var = NULL, by = NULL, type = 1L)
}
\arguments{
\item{dt}{a data.table}

\item{var}{name(s) of variable(s) with atomic values ; if 'var' is omitted, all variables not in 'by' are selected}

\item{by}{name(s) of variable(s) which determines groups (optional) ; if 'by' is ommitted, dt is considered as one group}

\item{type}{specifies type of filling : 0 constant 1 LOCF, 2 NOCB, 3 LOCF then NOCB. See note below.}

\item{fill}{when type=0, NA is replaced by this 'fill' value. For type = 1, 2 or 3, 'fill' is unused.}

\item{inplace}{when inplace = TRUE, na_fill is compute inplace}
}
\value{
a list with item for each var
}
\description{
Replace NA by constant, previous or next value, optionally by group
}
\note{
When \code{type = 0}, all NA values are replaced by the 'fill' value. Ex with 'fill = 0' : \code{c(NA, 1, NA, NA, 2, NA, 3, NA)}
gives \code{c(0, 1, 0, 0, 2, 0, 3, 0)}

When \code{type = 1}, LOCF = Last Observation Carry Forward. Ex : \code{c(NA, 1, NA, NA, 2, NA, 3, NA)} gives \code{c(NA, 1, 1, 1, 2, 2, 3, 3)}

When \code{type = 2}, NOCB = Next Observation Carry Backward. Ex : \code{c(NA, 1, NA, NA, 2, NA, 3, NA)} gives \code{c(1, 1, 2, 2, 2, 3, 3, NA)}

When \code{type = 3}, LOCF then NOCB. Ex : \code{c(NA, 1, NA, NA, 2, NA, 3, NA)} gives \code{c(1, 1, 1, 1, 2, 2, 3, 3)}.

When \code{type = 4}, NOCB then LOCF. Ex : \code{c(NA, 1, NA, NA, 2, NA, 3, NA)} gives \code{c(1, 1, 2, 2, 2, 3, 3, 3)}.
}
\examples{
library(data.table)
ngrp <- 18000
nbygrp <- 50
dt <- data.table(
  id=rep(1:ngrp, each=nbygrp),
  a=sample(1:100, nbygrp * ngrp, replace=TRUE),
  b=sample(LETTERS, nbygrp * ngrp, replace=TRUE),
  c=sample((1:100 + 0.5), nbygrp * ngrp, replace=TRUE),
  d=sample(1:100, nbygrp * ngrp, replace=TRUE),
  e=sample(1:100, nbygrp * ngrp, replace=TRUE)
)
dt[(1:.N \%\% 2 == 0), a:=NA]
dt[(1:.N \%\% 3 == 0), b:=NA]
dt[(1:.N \%\% 3 == 0), c:=NA]
dt[(1:.N \%\% 4 == 0), d:=NA]
dt[(1:.N \%\% 5 == 0), e:=NA]

vars = c("a", "c", "d", "e")

dt1 <- copy(dt)
na_fill_by(dt1, var=vars, by="id", inplace=TRUE)
dt2 <- copy(dt)
dt2[, (vars) := na_fill_by(.SD, vars, "id")]
dt3 <- copy(dt)
dt3[, na_fill_by(.SD, vars, "id", inplace=TRUE)]
}
