% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tsv.R
\name{write_tsv}
\alias{write_tsv}
\alias{write_tsv.data.table}
\title{TSV writer}
\usage{
write_tsv(x, file, ...)
}
\arguments{
\item{x}{An R object.}

\item{file}{Output file name.}

\item{...}{Additional arguments to be passed to methods.}
}
\description{
As fwrite, but defaults to TSV instead of CSV.
}
\examples{
library(data.table)
Mtcars <- as.data.table(mtcars, keep.rownames = TRUE)
write_tsv(Mtcars, "mtcars.tsv")
}
