% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tsv.R
\name{write_tsv}
\alias{write_tsv}
\alias{fwrite_tsv}
\alias{write_tsv.data.table}
\title{TSV writer}
\usage{
write_tsv(x, file, ...)
}
\arguments{
\item{x}{An \R object (e.g., a \code{\link{data.table}}).}

\item{file}{Output file name. File extensions other than \code{.tsv}, \code{.tab},
and \code{.txt} result in a warning.}

\item{...}{Additional arguments to be passed to methods
(e.g., to \code{\link{fwrite}}).}
}
\description{
As \code{\link{fwrite}}, but tab-separated (TSV) instead of the default
comma-separated (CSV).
}
\examples{
library(data.table)
library(dtutils)
mtcars_dt <- as.data.table(mtcars, keep.rownames = TRUE)
write_tsv(mtcars_dt, "mtcars.tsv")
}
